fix message token expired shown once

+++ src/utils/axiosInstance.js
@@ -7,20 +7,20 @@ const axiosInstance = axios.create({
     }
 });
 
-// Interceptor למקרה של טוקן פג תוקף
+let sessionHandled = false;
+
 axiosInstance.interceptors.response.use(
-    response => response,  // החזרת תוצאות מוצלחות
+    response => response,
     async (error) => {
-        if (error.response && (error.response.status === 422 || error.response.status === 401)) {
-            // טוקן פג תוקף או לא תקין
-            console.log("Token expired or invalid. Redirecting to login page.");
+        if ((error.response && (error.response.status === 422 || error.response.status === 401)) && !sessionHandled) {
+            sessionHandled = true;
             alert("Your session has expired or your token is invalid. Please log in again.");
-            // מחיקת הטוקן והפנייה לדף התחברות
-            localStorage.clear();  // ניקוי כל ה-localStorage
-            window.location.href = '/login';  // הפנייה לדף התחברות
+            localStorage.clear();
+            window.location.href = '/login';
         }
         return Promise.reject(error);
     }
 );
+
 
 export default axiosInstance;
